Updated the main.py to use async.

+++ main.py
@@ -1,22 +1,36 @@
-
+from contextlib import asynccontextmanager
 from typing import Annotated
 
 from fastapi import Depends, FastAPI, Request, HTTPException, status
+from fastapi.exception_handlers import (
+    http_exception_handler,
+    request_validation_exception_handler
+)
 from fastapi.exceptions import RequestValidationError
-from fastapi.responses import JSONResponse
+
 from fastapi.staticfiles import StaticFiles
 from fastapi.templating import Jinja2Templates
-from starlette.exceptions import HTTPException as StarletteHTTPException
+
 from sqlalchemy import select
-from sqlalchemy.orm import Session
+from sqlalchemy.ext.asyncio import AsyncSession
+from sqlalchemy.orm import selectinload
+from starlette.exceptions import HTTPException as StarletteHTTPException
 
 import models
 from database import Base, engine, get_db
 from schemas import PostCreate, PostResponse, PostUpdate, UserCreate, UserResponse, UserUpdate
 
 
-# Create database tables from our models
-Base.metadata.create_all(bind=engine)
+# Create database tables from our models # this is now the modern way.
+@asynccontextmanager
+async def lifespan(_app: FastAPI):
+    # startup
+    async with engine.begin() as conn:
+        await conn.run_sync(Base.metadata.create_all)
+    yield  # where the aplication works.
+    # Shutdown
+    await engine.dispose()
+
 
 # Create FastAPI app
 app = FastAPI()
@@ -40,12 +54,12 @@ templates = Jinja2Templates(directory="templates")
 # Homepage: / or /posts
 @app.get("/", include_in_schema=False, name="home")
 @app.get("/posts", include_in_schema=False, name="posts")
-def home(
+async def home(
     request: Request,
-    db: Annotated[Session, Depends(get_db)]
+    db: Annotated[AsyncSession, Depends(get_db)]
 ):
     # Get all posts from the database
-    result = db.execute(select(models.Post))
+    result = await db.execute(select(models.Post).options(selectinload(models.Post.author)))
     posts = result.scalars().all()
 
     # Render home.html and send posts to the template
@@ -61,14 +75,15 @@ def home(
 
 # Display one post: /posts/1
 @app.get("/posts/{post_id}", include_in_schema=False, name="get_post")
-def get_post(
+async def get_post(
     request: Request,
     post_id: int,
-    db: Annotated[Session, Depends(get_db)]
+    db: Annotated[AsyncSession, Depends(get_db)]
 ):
     # Find the post by ID
-    result = db.execute(
-        select(models.Post).where(models.Post.id == post_id)
+    result = await db.execute(
+        select(models.Post).options(selectinload(
+            models.Post.author)).where(models.Post.id == post_id)
     )
     post = result.scalars().first()
 
@@ -82,7 +97,7 @@ def get_post(
             "post.html",
             {
                 "post": post,
-                "title": title 
+                "title": title
             }
         )
 
@@ -96,13 +111,13 @@ def get_post(
 # Display all posts belonging to one user
 # Example: /users/1/posts
 @app.get("/users/{user_id}/posts", include_in_schema=False, name="user_posts")
-def user_posts(
+async def user_posts(
     request: Request,
     user_id: int,
-    db: Annotated[Session, Depends(get_db)]
+    db: Annotated[AsyncSession, Depends(get_db)]
 ):
     # Find the user
-    result = db.execute(
+    result = await db.execute(
         select(models.User).where(models.User.id == user_id)
     )
     user = result.scalars().first()
@@ -114,8 +129,9 @@ def user_posts(
         )
 
     # Get all posts belonging to the user
-    result = db.execute(
-        select(models.Post).where(models.Post.user_id == user_id)
+    result = await db.execute(
+        select(models.Post).options(selectinload(models.Post.author)).where(
+            models.Post.user_id == user_id)
     )
     posts = result.scalars().all()
 
@@ -141,12 +157,12 @@ def user_posts(
     response_model=UserResponse,
     status_code=status.HTTP_201_CREATED,
 )
-def create_user(
+async def create_user(
     user: UserCreate,
-    db: Annotated[Session, Depends(get_db)]
+    db: Annotated[AsyncSession, Depends(get_db)]
 ):
     # Check if username already exists
-    result = db.execute(
+    result = await db.execute(
         select(models.User).where(
             models.User.username == user.username
         )
@@ -160,7 +176,7 @@ def create_user(
         )
 
     # Check if email already exists
-    result = db.execute(
+    result = await db.execute(
         select(models.User).where(
             models.User.email == user.email
         )
@@ -182,8 +198,8 @@ def create_user(
 
     # Save user to database
     db.add(new_user)
-    db.commit()
-    db.refresh(new_user)
+    await db.commit()
+    await db.refresh(new_user)
 
     return new_user
 
@@ -191,12 +207,12 @@ def create_user(
 # Get one user
 # Example: GET /api/users/1
 @app.get("/api/users/{user_id}", response_model=UserResponse)
-def api_get_user(
+async def api_get_user(
     user_id: int,
-    db: Annotated[Session, Depends(get_db)]
+    db: Annotated[AsyncSession, Depends(get_db)]
 ):
     # Find user by ID
-    result = db.execute(
+    result = await db.execute(
         select(models.User).where(models.User.id == user_id)
     )
     user = result.scalars().first()
@@ -216,12 +232,12 @@ def api_get_user(
     "/api/users/{user_id}/posts",
     response_model=list[PostResponse]
 )
-def get_user_posts(
+async def get_user_posts(
     user_id: int,
-    db: Annotated[Session, Depends(get_db)]
+    db: Annotated[AsyncSession, Depends(get_db)]
 ):
     # Check that the user exists
-    result = db.execute(
+    result = await db.execute(
         select(models.User).where(models.User.id == user_id)
     )
     user = result.scalars().first()
@@ -233,43 +249,46 @@ def get_user_posts(
         )
 
     # Get user's posts
-    result = db.execute(
-        select(models.Post).where(models.Post.user_id == user_id)
+    result = await db.execute(
+        select(models.Post).options(selectinload(models.Post.author)).where(
+            models.Post.user_id == user_id)
     )
     posts = result.scalars().all()
 
     return posts
 
-#patch for user.
+# patch for user.
+
+
 @app.patch("/api/users/{post_id}", response_model=UserResponse)
-def api_update_user_partial(
+async def api_update_user_partial(
     user_id: int,
     user_update: UserUpdate,
-    db: Annotated[Session, Depends(get_db)]
+    db: Annotated[AsyncSession, Depends(get_db)]
 ):
     # Find post by ID
-    result = db.execute(
+    result = await db.execute(
         select(models.User).where(models.User.id == user_id))
     user = result.scalars().first()
 
     if not user:
         raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail="User not found"
-            )
-    if user_update.username is not None and user_update.username!= user.username:
-        result = db.execute(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="User not found"
+        )
+    if user_update.username is not None and user_update.username != user.username:
+        result = await db.execute(
             select(models.User).where(models.User.username == user_update.username),)
         existing_user = result.scalars().first()
         if existing_user:
             raise HTTPException(
-                    status_code=status.HTTP_400_BAD_REQUEST,
-                    detail="Username already exists",
+                status_code=status.HTTP_400_BAD_REQUEST,
+                detail="Username already exists",
             )
-        if user_update.email is not None and user_update.email!= user.email:
-                result = db.execute(
-                    select(models.User).where(models.User.username == user_update.username),)
-                existing_email = result.scalars().first()
+        if user_update.email is not None and user_update.email != user.email:
+            result = await db.execute(
+                select(models.User).where(models.User.username == user_update.username),)
+            existing_email = result.scalars().first()
         if existing_email:
             raise HTTPException(
                 status_code=status.HTTP_400_BAD_REQUEST,
@@ -278,22 +297,22 @@ def api_update_user_partial(
     if user_update.username is not None:
         user.username = user_update.username
     if user_update.email is not None:
-            user.email = user_update.email   
+        user.email = user_update.email
     if user_update.image_file is not None:
-            user.image_file = user_update.image_file
-    
-    
-    db.commit()
-    db.refresh(user)
+        user.image_file = user_update.image_file
+
+    await db.commit()
+    await db.refresh(user)
     return user
 
+
 @app.delete("/api/users/{user_id}", status_code=status.HTTP_204_NO_CONTENT)
-def api_delete_user(
+async def api_delete_user(
     user_id: int,
-    db: Annotated[Session, Depends(get_db)]
+    db: Annotated[AsyncSession, Depends(get_db)]
 ):
     # Find user by ID
-    result = db.execute(
+    result = await db.execute(
         select(models.User).where(models.User.id == user_id)
     )
     user = result.scalars().first()
@@ -302,10 +321,9 @@ def api_delete_user(
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail="User not found"
-        ) 
-    db.delete(user)
-    db.commit()
-
+        )
+    await db.delete(user)
+    await db.commit()
 
 
 # ============================================================
@@ -314,11 +332,11 @@ def api_delete_user(
 
 # Get all posts
 @app.get("/api/posts", response_model=list[PostResponse])
-def api_get_posts(
-    db: Annotated[Session, Depends(get_db)]
+async def api_get_posts(
+    db: Annotated[AsyncSession, Depends(get_db)]
 ):
     # Get all posts from database
-    result = db.execute(select(models.Post))
+    result = await db.execute(select(models.Post).options(selectinload(models.Post.author)),)
     posts = result.scalars().all()
 
     return posts
@@ -330,13 +348,14 @@ def api_get_posts(
     response_model=PostResponse,
     status_code=status.HTTP_201_CREATED,
 )
-def create_post(
+async def create_post(
     post: PostCreate,
-    db: Annotated[Session, Depends(get_db)]
+    db: Annotated[AsyncSession, Depends(get_db)]
 ):
     # Make sure the user exists
-    result = db.execute(
-        select(models.User).where(models.User.id == post.user_id)
+    result = await db.execute(
+        select(models.User).where(
+            models.User.id == post.user_id)
     )
     user = result.scalars().first()
 
@@ -355,8 +374,8 @@ def create_post(
 
     # Save post to database
     db.add(new_post)
-    db.commit()
-    db.refresh(new_post)
+    await db.commit()
+    await db.refresh(new_post, attribute_names=["author"])
 
     return new_post
 
@@ -364,13 +383,14 @@ def create_post(
 # Get one post as JSON
 # Example: GET /api/posts/1
 @app.get("/api/posts/{post_id}", response_model=PostResponse)
-def api_get_post(
+async def api_get_post(
     post_id: int,
-    db: Annotated[Session, Depends(get_db)]
+    db: Annotated[AsyncSession, Depends(get_db)]
 ):
     # Find post by ID
-    result = db.execute(
-        select(models.Post).where(models.Post.id == post_id)
+    result = await db.execute(
+        select(models.Post).options(selectinload(
+            models.Post.author)).where(models.Post.id == post_id)
     )
     post = result.scalars().first()
 
@@ -381,77 +401,19 @@ def api_get_post(
         status_code=status.HTTP_404_NOT_FOUND,
         detail="Post not found"
     )
-##put endpoint.
+
+
+# put endpoint.
 @app.put("/api/posts/{post_id}", response_model=PostResponse)
-def api_update_post_full(
+async def api_update_post_full(
     post_id: int,
     post_data: PostCreate,
-    db: Annotated[Session, Depends(get_db)]
+    db: Annotated[AsyncSession, Depends(get_db)]
 ):
     # Find post by ID
-    result = db.execute(
-        select(models.Post).where(models.Post.id == post_id)
-    )
-    post = result.scalars().first()
-
-    if not post:
-        raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail="Post not found"
-            )
-    if post_data.user_id != post.user_id:
-         result = db.execute(select(models.User).where(models.User.id == post_data.user_id))
-         user = result.scalars().first()   
-        #CHECKING IF USER.
-         if not user:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail="User not found"
-            )
-    post.title = post_data.title  
-    post.content = post_data.content
-    post.user_id = post_data.user_id    
-
-    db.commit()
-    db.refresh(post)
-    return post   
-
-#partial update - that is patch.
-@app.patch("/api/posts/{post_id}", response_model=PostResponse)
-def api_update_post_partial(
-    post_id: int,
-    post_data: PostUpdate,
-    db: Annotated[Session, Depends(get_db)]
-):
-    # Find post by ID
-    result = db.execute(
-        select(models.Post).where(models.Post.id == post_id)
-    )
-    post = result.scalars().first()
-
-    if not post:
-        raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail="Post not found"
-            )
-
-    update_data = post_data.model_dump(exclude_unset=True)
-    for field, value in update_data.items():
-        setattr(post, field, value)
-    
-    db.commit()
-    db.refresh(post)
-    return post  
-
-#api end point for deleting 
-@app.delete("/api/posts/{post_id}", status_code=status.HTTP_204_NO_CONTENT)
-def api_delete_post(
-    post_id: int,
-    db: Annotated[Session, Depends(get_db)]
-):
-    # Find post by ID
-    result = db.execute(
-        select(models.Post).where(models.Post.id == post_id)
+    result = await db.execute(
+        select(models.Post).options(selectinload(
+            models.Post.author)).where(models.Post.id == post_id)
     )
     post = result.scalars().first()
 
@@ -459,10 +421,77 @@ def api_delete_post(
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Post not found"
-        ) 
-    db.delete(post)
-    db.commit()
-    
+        )
+    if post_data.user_id != post.user_id:
+        result = await db.execute(select(models.User).where(
+            models.User.id == post_data.user_id))
+        user = result.scalars().first()
+        # CHECKING IF USER.
+        if not user:
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND,
+                detail="User not found"
+            )
+    post.title = post_data.title
+    post.content = post_data.content
+    post.user_id = post_data.user_id
+
+    await db.commit()
+    await db.refresh(post, attribute_names=["author"])
+    return post
+
+# partial update - that is patch.
+
+
+@app.patch("/api/posts/{post_id}", response_model=PostResponse)
+async def api_update_post_partial(
+    post_id: int,
+    post_data: PostUpdate,
+    db: Annotated[AsyncSession, Depends(get_db)]
+):
+    # Find post by ID
+    result = await db.execute(
+        select(models.Post).options(selectinload(
+            models.Post.author)).where(models.Post.id == post_id)
+    )
+    post = result.scalars().first()
+
+    if not post:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="Post not found"
+        )
+
+    update_data = post_data.model_dump(exclude_unset=True)
+    for field, value in update_data.items():
+        setattr(post, field, value)
+
+    await db.commit()
+    await db.refresh(post, attribute_names=["author"])
+    return post
+
+# api end point for deleting
+
+
+@app.delete("/api/posts/{post_id}", status_code=status.HTTP_204_NO_CONTENT)
+async def api_delete_post(
+    post_id: int,
+    db: Annotated[AsyncSession, Depends(get_db)]
+):
+    # Find post by ID
+    result = await db.execute(
+        select(models.Post).options(selectinload(
+            models.Post.author)).where(models.Post.id == post_id)
+    )
+    post = result.scalars().first()
+
+    if not post:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="Post not found"
+        )
+    await db.delete(post)
+    await db.commit()
 
 
 # ============================================================
@@ -471,23 +500,21 @@ def api_delete_post(
 
 # Handles HTTP errors such as 404
 @app.exception_handler(StarletteHTTPException)
-def general_http_exception_handler(
+async def general_http_exception_handler(
     request: Request,
     exception: StarletteHTTPException
 ):
-    # Get error message
+
+    # API errors return JSON
+    if request.url.path.startswith("/api"):
+        return await http_exception_handler(request, exception)
+
+        # Get error message
     message = (
         exception.detail
         if exception.detail
         else "An error occurred. Please check your request and try again."
     )
-
-    # API errors return JSON
-    if request.url.path.startswith("/api"):
-        return JSONResponse(
-            status_code=exception.status_code,
-            content={"detail": message}
-        )
 
     # Website errors return error.html
     return templates.TemplateResponse(
@@ -505,16 +532,13 @@ def general_http_exception_handler(
 # Handles validation errors
 # Example: /posts/abc when post_id expects an integer
 @app.exception_handler(RequestValidationError)
-def validation_exception_handler(
+async def validation_exception_handler(
     request: Request,
     exception: RequestValidationError
 ):
     # API validation errors return JSON
     if request.url.path.startswith("/api"):
-        return JSONResponse(
-            status_code=status.HTTP_422_UNPROCESSABLE_CONTENT,
-            content={"detail": exception.errors()}
-        )
+        return await request_validation_exception_handler(request, exception)
 
     # Website validation errors return error.html
     return templates.TemplateResponse(
